Add working sign up modal
